+ rc: initial import from untracked
this is from 6 years ago when experimenting with taskwarrior, shelved
because there was no decent mobile client at the time.  we'll try to use
it again and figure because we need something more sophisticated than a
text file at the moment.  there's a few mobile options, but none of them
look good.  we'll try to figure that part out later...