- rc: report: list: sort: entry date, then urgency, then priority
this seemed to be a lex sort before setting it explicitly.  at first
made it urgency but really this list isn't used normally, only 'next'
would be.  when we're looking at this last, we probably want old ones to
come up first.  we'll see how this works out in time.