/ taskpy: taskw.get_task() has other type so always use .filter_tasks()
for the types of task search that would always return one match --
integer task id, and uuid -- we just used taskw.get_task() because it
should always return one match anyways.  but that makes it return a
different type ie (idnum, task) rather than list of tasks, so we
couldn't use the same return value or would have had to extract the task
and put it into a list.

rather than do that we just changed them all to use filter_tasks() so
they can keep the same logic, and then we added a safety to check that
there is indeed only one task returned.  it should never happen that
multiple tasks are returned for these search types -- if any are
returned, only one should be returned -- but we'll leave the safety in
there in case some other assumption is broken elsewhere