= taskpy: _taskget: call it 'tasks' instead of 'task' since is list
these are returning multiple tasks as a list, not just one.  ok so the
first two from id and full-uuid are only returning one task, but all
others can return multiple, and even the first two return a list.  so
they all are returning "tasks" not "a task."  change variable name
accordingly