+ taskpy: taskdo: make it handle no args, for timew continue
giving no args will allow to continue last interval (ie @1, ie
timew.cont(1)), otherwise we switch to the specified one (as
timew.start() does).

we also check whether we should be allowed to do that, for example if
that task was completed in taskwarrior, we will not be able to work on
it any longer.