+ taskpy: _taskone: give it an abort param to allow to keep going
normally it's calling bomb() when it can't find exactly one matching
task, but sometimes we want this, example when we have a "fake" timew
tag with no corresponding task -- like time/todo tag -- as the last one,
and we want to resume it.  we'll get the fql of @1 and look up that fql,
but find no task.  we need to be able to keep going and get to
timew.continue() because this is [almost?] always what we want to do
at that point