/ taskpy: _taskdo: allow keep going if @1 fql not in taskw
we use some simulated tags like time/todo in timewarrior to account for
time we're doing "metawork" with time management, ie manipulating tasks
themselves, but it's not to be logged against any specific task.
however we don't want to lose this time so we use timew tags that don't
have any taskw task.  so far we have reserved the time/ tag namespace
for this

if user runs `taskdo` alone with no args, it will need to look up last
task, which could be such a "fake" task and we need to handle that and
take it to mean "just run timew continue"