- rc: replace annotation text with only a count thereof in reports
if we're using annotations to track progress, they can get lengthy and
accumulate, taking up a lot of space.  we don't want them to display in
reports, which are intended as an overview only.  however we still want
to see that annotations are present, which we accomplish using .count
modifier (see "task columns")