/ taskget: use .contains rather than '=' for description substring
could not figure out why 'taskids the' returned only one match (the one
where description *started with* 'the'.  the reason is actually because
of this from the documentation:

    Strings compare equal if the left operand starts with the right
    operand.  The == operator tests for exact equality.

otoh, .contains is really a substring match, which is what we really
wanted in the first place, so this was a bug.