- taskids: separate multiple ids by spaces rather than newlines
when we output, should probably all be one line since it's supposed to
be in a subshell anyways.  really the idiom "task $(taskids foo)" does
work anyways in practice when output is on separate lines, but just to
make clear that we're constructing a line of whitespace-separated
arguments we'll do it this way.  plus, if we want to do a quick check to
see how many ids are generated, we don't necessarily want multiple
lines; we have eg "taskfqls" for this