+ taskget: implement -a/--all to accumulate different match types
normally taskget is only returning the first "kind" of match and doesn't
keep looking for other kinds.  for example if it matches on a project
substring, it won't keep looking for substrings in the description.

now we add a -a/--all argument that will keep looking and build up a
cumulative list of matches from all the match sources, rather than
giving only the list from the first type that matches

this also implicates taskids and taskfqls, which were using taskget.
now we can take the -a there also and output multiple matches

this is helpful if we're looking for eg "all tasks that have ansible
anywhere in description, project, etc"