+ taskget: deduplicate list of matches now that we have --all
when passing --all to taskids/taskfql/taskget we can end up with a lot
of duplicates that match by different ways especially things present
both in description and project.  no need to return all those duplicates
and force the callers to dedup in multiple places, there should be no
need ever for taskget to return duplicates anyways so just make it do
that

was not successful immediately thinking of a clever single-line
statement to do this same thing.  maybe in addition to a possible
comprehension, also there might be a way using functools.reduce()