- addargs: new argparse wrapper for positional arg with nargs='*'
the default for nargs behavior is to treat as taking exactly one arg,
but doesn't return a list like it would if we did nargs=1.  we want to
support multiple specified arguments and this is done with nargs='*'.
we pass in this value to the add_argument() call when wrapped using
addargs() (ie plural with an 's') rather than addarg() (singular).

the need for this, ultimately, is to allow taskget to take multiple
arguments