+ main: handle broken pipe during subprogram execution
eg if output is piped to pager and we terminate it before the pagers reads
the whole task.py output buffer.  no reason this should cause a crash...

note that it should return a bad exit, but sys.exit(number) apparently
isn't working, it's always returning 0.  not sure why this is;
sys.exit(n) is raising SystemExit(n) and that should work fine.  not
sure if already being in an exception handler has something to do with
it? we can figure this out later it's not really important