- taskget: reimplement dedup by using set of hashable dicts
sets are the most natural way to deduplicate (imo), without needing a
second pass on the compiled filter results.  but we cannot add a list of
dicts to a set, because set inserts use hash of the item to determine if
already present, and dicts are not hashable: their __hash__ members
simply return None.

however, in our case, all these dicts are tasks returned by python-taskw
library's TaskWarrior class filter_tasks method, which means they all,
for certain, have a 'uuid' key.  this is both (1) unique and (2)
hashable.  So we change the running task list from a list of
dictionaries to a set of these UUIDHashableDicts we have made which
provides a suitable __hash__ method, and then change all the
addition-assignment operators to set updates.

lastly, we can remove the deduplication pass at the end of the loop
altogether.