/ taskget: don't exit if uuid does not match, as we need more passes
there are other arguments to process; just because this one didn't
match, we don't need to stop altogether (as long as --all was specified)

however we *can* stop processing this particular argument, because we
know it's a full uuid, which will never match.  this does mean we won't
end up looking up the uuid in description or notes though, which may or
may not be a problem.  if this ever comes up we'll have to look into it
later