+ optparse: centralize option parsing so different cmds have their own
we use some things like _taskget() or _taskone() internally, but these
do their own option processing, unrelated to option processing in the
"wrapper" command invoked from the shell (which has more
command-specific options).  instead of trying to work out argument
processing and passing unfinished arguments around, we'll do all option
parsing in a separate function that always makes a new ArgumentParser
every time, and keeps track of the "known" and "unrecognized" arguments,
storing them for the next parser run (which may understand these
arguments)

the Namespace structure at the end is basically a union of all these
separate option passes, but the first setter wins, because that will eat
it from the list of arguments that can be processed later.  This is not
really satisfactory, but works for now, in all the use cases we care
about, and enables each command to set its own arguments (the other
option is a global argument prcessing done in __main__ section, or
something like that.

TODO last setter should win