/ taskget: use empty args when not provided so search can default
needs to have a set with a nothing in it, rather than being a nothing
itself.  because in the main loop through taskargs, we have a case to
handle where the arg exists but is None.  and that's what we want to
indicate here

the program logic changed, this used to not be necessary but something
has changed around no args in the last so many commits where this now
needs to have an empty list given.  in any case it seems to work for
most things in rudimentary tests, so we'll keep it for now and fix any
bugs later