+ taskget: idonly: new kwarg: search no further than id uuid label fql
we want to be able to treat some things that call taskget as wanting to
search through everything in taskget (following the taskget algorithm)
*up through* an fql, after which we start doing text searches in
different fields and getting more desperate for a match.

with some commands (ie users of _taskget()) we may want to just return
if we haven't found anything, because what we would have wanted to match
on (some form of unique identifier for the task, currently there are
four: [locally-unique] id, uuid, label, fql) didn't match, we already
know looking any further will only find bad matches.  so far we are
treating 'taskid' and 'taskuuid' this way: it's expected that they only
get passed one of those simple things (id, fql, label, uuid)

(next commit will implement this).

this makes sense also to be able to get the uuid of a purely-timew task
(such as "time/todo") and have it correctly give a bogus uuid as output,
when otherwise it could continue on and match the
descriptions/annotations of any task