+ taskday: new option to print status chars after each entry
we do a lookup and find their status, to get the char and display it
after their label (or fql, depending on which options have been used).
doing this can be expensive, so for example 'taskday -s 999' will take a
while because that's a lot of separate calls to 'task export'
internally.  but that's ok for the moment because the typical use case
is eg 'taskweek -s' which doesn't take enormously long, we will have to
optimize this later though, if we have a 'taskmonth -s' later this could
take unacceptably long

status chars are for completed, pending, deleted, not a taskwarrior task
(eg just a timew pseudo-task with a label there but no corresponding
label here)