- taskget: use exact matching on label part of an fql
in taskget() and taskfql() we match an fql based on project and label
matching but the default is not exact match, it's whole-substring match
(ie '.contains' iirc).  this can result in multiple matches when we give
a full and complete, unique label (when another task uses it as a
leading substring for its own label).  this happens often in practice
because of the limitation that labels are unique