- taskget: change noarg default to be current task so callers neednt
previously, the default when no argument was supplied to taskget() was
to return "all tasks," which isn't really a useful thing to request, or
at least there hasn't been a need so far; if we wanted to select all
tasks we could always do that using 'task' itself and export it or
whatever... it would be a rare / unusual special case where special
processing would be called for anyways

instead we can make the default with noargs be to return the
*current* task -- ie the one being worked on now.  this saves us the
trouble of doing it in several different callers which are wanting
to deal with the current task when they get no argument (such as
tasknotes, taskdo, taskid, or taskget itself)

the purpose of these tools is to find and work with subsets
of tasks, not all tasks.  if ever a use case comes up we can consider
adding a flag to list everything without an argument, that might
actually be a good idea to have... because there could be some scenarios
where taskget would be useful.  taskget --dumpall --json sounds like an
interesting use case, but maybe that would be the same as "task export"