+ taskdo: give ability to pass cli args to taskget
pass and let _taskget() process the args, since most of the ones we
might want to supply in using 'taskdo' would be used to change said
function's behavior.

we can't check if arg is None anymore, that will be done by _taskget()
now, and it will return the current task if no arg was supplied.  we
can make further changes as a result of this, since the function is
currently checking if task is None, but that's vestigial, the code path
won't be taken, and won't hurt anything, we'll remove it later