- taskdo: simplify now that taskget is handling no args itself
now that _taskget() has been changed to give the current task when no
args are given, we know we will always get a response from _taskget(),
so we don't have to handle empty task ourselves anymore, we can rely on
a task being returned from _taskget() and remove some logic.  the result
is quite a bit simpler.

in the future we should find a way to re-use the result of _tasknow()
after it gets looked up when no arg to _taskget(), but at the moment
this is too far removed and we have no way to do this simply.  possibly
another good plan would be to have _tasknow() cache its result, which
should be fine to be the same throughout the script.  at least right now
there are no internal use cases which consult this value twice.  we
would already know if we changed the task because we did it ourselves,
so this seems unlikely to be needed, but we'll see