+ tasknotes: ability to output multiple tasks given as args on cli
now that we're passing args to taskget, we can also pass several args,
the user could be requesting notes on more than one thing.  not likely
to be very frequently used, but it could be desired in some cases and is
easy enough to support by just doing a loop.

oh, one obvious use case is if some flag (to be implemented later) is
given to request only one-liner descriptions, it would form a concise
task list of the search results, which you might want to supply multiple
args to when making a list of something for copy/paste elsewhere.  turns
out there's already a task e6c0ef4c-ef2b-4320-860f-dcbb781d1197 for this