/ taskdo: make sure task selected to do matches unambiguously
when selecting a task to start, we want to make sure the argument we
provide is unambiguous, else abort.  we want to prevent accidents when
tracking time and starting/stopping projects.  without aborting, it
would just use the first one if, say, we had a taskone and a taskonetwo,
and we would never be the wiser that the wrong/unintended one was
selected