- tasknow: cache result as some paths call _tasknow() twice
so far taskdo() with no arguments, but potentially others will end up
calling tasknow twice.  we can mitigate this by caching the result.  it
should always be ok because even if -- in the future, because it doesn't
today -- the script changes the current task and then does something
else, that's fine because the script will be the one that did it and
won't need to request again the current task since it requested itself