+ taskget: add -x/--exact flag for .is and make .has default again
this was the default before and we removed it in commit 2ef1f62, because
we commonly have tasks with common substrings.  however we still want
the ability to simply express tasks with eg our 'taskl' alias which just
runs 'taskid' with flags, and it will not match unless exact because we
used .is, so now we have to spell out everything, which is inconvenient.

instead add a -x or --exact for anything that ends up calling _taskget()
to use .is, otherwise use .has like before (which we now revert to be
the default)