/ taskone: recognize taskget failure when --zero is in effect
with --zero we still get one result from _taskget(), so we cannot use
result count alone to determine failure.  if count is one, we must look
to see if it's a failure uuid.  thanks to last commit 89877d8 where we
changed the fail uuids to enums which are bitmask lengths and now we can
now utilize the new function isfailuuid() that tests if these uuids
indicate errors.  this works around not being able to pass the error
code, only the status, and opens up for more sophisticated
interpretation of and/or more of the error uuids later