/ timewtags: fix output not to use --zero by default
unless cli args override it, we should not do this, because the context
it's used asks for tags, so supplying a uuid could make timewarrior end
up putting uuids as tags if used in a subshell.  better to just return
nothing, maybe this could result in tags getting erased though.  we'll
at least be returning bad exit code on failure due to abort=True being
the default for taskone()