/ onmodify: read all stdin ahead of time so pdb can use fds
pdb needs to interact with the user on stdio.  so we have to read all
the stdin and then use the buffer we read it into rather than sys.stdin
directly.  pdb will open 'sys.stdin' directly (in absence of stdin=
argument to constructor) and use it, so we have to use a separate
variable to store the lines and overwrite sys.stdin