/ taskfql: fix onmodify hook by returning None if fql malformed
during taskadd, the old task will be one without a project or label, and
that has to work, so fql has to return nothing and we'll bypass further
hook processing

this broke in 2c56ea9 when we stopped gating checks on a label with
existence of a project.  but really the hook should just not run if both
are present, rather than thinking this is an error.  all control for new
tasks using task.py scripts should be happening through taskadd, so
there should never be a malformed one created unless it was
intentionally bypassing the task.py scripts anyways