+ taskday: cli flags to select pending, timew-only, completed tasks
make the default not to include these because they're typically not
useful, but then let one or both be turned back on

notably we still don't have a way to select just one or the other and
not include other pending tasks

also notably we don't have a way to select just active (started) tasks,
even though come to think of it that information is available right at
the place we filter tasks.  actually that's a good optimization that
should be a in follow-up commit

anyways so what this patch does is change the function that did taskw
lookup (since we only had timew data so far) for purposes of figuring
out the status char (if --status flag active), and changed it to also
return the status itself (or None if it doesn't exist, such as with
timew-only tasks; these are then removed from the list by a filter on
truthiness).  then it runs unconditionally now, whether --status was
active or not, because we'll also now use the returned status in the
caller -- ultimately, fql_among_tags()) to filter which tasks to return
at all, based on cli args to choose which to select among timew-only
(--timetasks), completed (--done), pending (default), and all of the
above (--alltasks)