/ taskget: make uuid match behavior same full vs initial uuid
behavior was different between initial-substring uuid and full uuid, for
former it would keep going on no match, whereas we didn't want this
behavior for latter.  it's really the same situation -- if the uuid
matches we want to stop, if it doesn't but looks like one then it's
under control of args.idstrings if we stop -- so it should behave the
same

doing this makes all three (id, uuid, initial-uuid) the same really, so
maybe we can factor these out into a common routine and call it? TODO