- rc: open: change to minimum pattern after testing
the source says it already anchors using ^ before attempting the pattern

as for '$', even without it, for some reason it won't match if there's
anything after the pattern, as though a '$' exists but there's no '$'
and the code doesn't show it adding one, so this doesn't really make
sense.  even adding '.*' at the end doesn't work.  'file:' must be on a
line by its own, 'file: this is a test' will not work no matter what
pattern is used.  not clear why but give up trying to figure it out.
we'll just have to keep it by itself on a line for now

in meantime change to quoted forms, which seem to work still and are
safer

really we don't need taskopen because it's huge and we only use a couple
things that would be easy to reproduce ourselves.  see task 18829637