+ taskday: implement a held state, which hides +WAITING by default
usually we're using taskday (and its different-timespan aliases) to find
things to work on, so we want to ignore held tasks (what we call those
that are waiting).  there is a virtual state, but we cannot access it
using ralphbean's python-taskw because that one can only really do
foo:bar on the command line and there is no way to access virtual tags
that way (but see taskwarrior issue 2908)

instead we found a way that we can access waiting state using ':' syntax:

  "wait.after:now" can be used instead of +WAITING
  "wait.none:" can be used instead of -WAITING

so now we filter waiting tasks from both taskday and next reports (we
were doing that already with our fake 'hold' state, which went unused).
we can use this as a pseudo-"hold" state effectively

see also task eb595b9a