- tasklib: switch from old ralphbean taskw library to tasklib
should be otherwise identical, just uses different library.  this one
takes filter params as kwargs rather than a dict, so it needs to have
dots replaced with double underscores.  there are a few other quirks but
this was mostly straightforward.  the biggest issue was with tasks no
longer having a .get() method, which we could have made a PR for, but it
was easy enough because they made task __getitem__() return null, this
meant only for tasks that aren't even real at all, such as time/todo.
we have to handle it specially vs before.  mainly this only comes up
in get_status(), which required some juggling, but is more clear now.

ran some timings, it performs about the same as ralphbean.

advantage is, this software is somewhat maintained vs PRs never being
reviewed at all (we'll see, have two pending here so far, but just
submitted)

second advantage is, we can pass arbitrary expressions to '.filter()'
method, which we cannot with the one from ralphbean.  this lets us
finally access virtual tags, although we cannot do it using "native"
filter types, just have to provide another arg to .filter() method,
which should be totally fine anyways