/ taskfql: fix overzealous .get() replacement in previous commits
when moving to tasklib, we no longer could use .get() because this
method is not implemented.  instead, access to keys that don't exist
return None, whihc is the reason to use .get() in the first place.
however, we were also doing it in __taskfql(), and actually it had
nothing to do with the task library, so this was just mistaken
replacement