- taskday: select_with_status: nee get_status, return with statchars
instead of pasting the status char onto the string depending on
args.status, caller should be doing this.  we just want to add the
information about status to the tasks and let caller decide what to do

this generalizes the status lookup function to one that returns pairs,
(fql, status), and then the caller can decide what it wants to do with
the extra info, rather than us deciding and pasting (and returning) a
buffer.

this also allows differentiation when the "extra" thing to paste is
actually an empty string.  this is impossible to tell once the string
has already been combined, whereas if we carry around as pairs, an empty
string is still a "thing" of its own we can keep track of and associate
with the fql (and decide to prepend/append to the string on print, etc)