/ taskday: fix saying last task is active even if pseudo-task
if pseudo-tasks (those in timew but not in taskw) are not displayed,
then the "active" indicator in front of the first (most recent) task
will be wrong.  it will look like that task is active, but really it's a
hidden one.  print % in this case, when the currect active task is
virtual.  this way the user will "know" that the "real" current task is
hidden and does not align with the shown first task.