/ dmain: fix use in subshell consuming unrelated stdin accidentally
there are issues with pdb and stdin, which we are already working around
in b4f548b, but this causes problems when running in a subshell pipeline
that's consuming stdio, because we come and wreck the descriptor by
reading from it.  really stdin is never for us, except in the case of
trigger.

without the fix:

    $ (echo 1; echo 2) |
      while read line
      do echo $(./timewtags $line)
      done

    omnius/ omnius/rc/ omnius/rc/alternatives +system

with the fix:

    $ (echo 1; echo 2) |
      while read line
      do echo $(timewtags $line)
       done

    omnius/ omnius/rc/ omnius/rc/alternatives +system
    adm/ adm/pkg/ adm/pkg/aptmove +look +system

so one can see, that we should not consume stdin, unless it's for us.
however, the real fix is not to test stdio readiness at all (with
select, as the existing code does) and just let that be used as normal.
pdb should really just use a different file descriptor than the stdio
ones, but we've been punting that becuase it becomes more complicated to
debug after doing that, and also means we lose readline in pdb if we
provide our own stream... so we'll only do this later if forced by some
other problem.