- timewtags: should pass exact match flag for this cli invocation
this is almost always used in a subshell and should match full labels or
projects, otherwise it will get multiple matches on things that have
same task initial substring (which are many) and expand to empty.
cannot think of a use for more fuzzy matching with timewtags.  can
always get it by using eg $(timewtags $(taskids ...)) to get access to
inexact matching