= taskget: taskfilter: comment that wait:none not same as -WAITING
these are different, in particular the waits that have already passed
will cause wait:none to fail match, even though the task is no longer
waiting.

we've chosen to audit for this out of band (in timechk) rather than fix
the filter, because removing the "wait" should be something explicit and
might cause user to take action, it should not happen in trigger or
automatically or something like that.  so we just audit and mail the
user and they can decide to remove 'wait' or make it a later time