- taskadd: parse_args: warn if description seems too short
code takes the last arg and uses it for description, OR concatenates
every arg after '--'

user may have forgotten to use one or the other.  so description length
check just ends up checking the last arg, which is always one word long
and thus quite short.  if we notice it's less than 20 chars, this is
probably what happened, so we warn the user