+ taskday: add proper knowledge and display of task wait states
taskday was using ^ as the statchar for waiting tasks, because taskget
was getting called with wait:none filter via taskone (because held not
requested, ie a normal run), and this won't find waiting tasks.  so
taskone will get a failure, which it equates to a virtual task.  there
is actually no way to distinguish at that point, so we have to remove
the filter, retrieve all of them and then compare the waiting field
ourselves

along the way this gives us the ability to know if they are waiting or
not, and if the wait period has already passed.  so we can add two new
virtual task states ('waiting' and 'waited') with corresponding status chars