/ taskget: tasklib uses _data member to keep the raw task data
otherwise the default for __repr__ is just to return description.  we
actually want a dump of all the members like before, and this is what
tasklib stores in the _data member.  the underscore means it's private,
sure, not sure about that.  maybe there should be a accessor method for
doing this in the library.  in the meantime we'll see if this every
breaks and figure it out then

eventually we want a json export too, actually tasklib has that
available already, whereas python-taskw did not