- dmain: name the FailMask enum, even though only used for repr()
works fine without the name, and seems this is only really used to have
a symbolic name when making a repr().  nonetheless it looks strange to
be empty.  it looks strange with it also, because the name of the thing
is the same as the string you're passing and seems redundant, but this
happens in a lot of places in python.

it would be a good idea if every object had an attribute that could be
accessed that said the name of the variable that pointed to it.  even
though multiple variables can point to it, the attribute retrieval
process could differentiate by maintaining a list of referers every time
a reference to it was initialized, and giving all users a "view" on the
self-referential object that includes its own name.  this way the object
could be passed around with another piece of information about itself,
without having to specify -- such as not needing this second string
parameter.  (the variable could be renamed, so there would need to be a
trigger to update the object's reference table)

now if you wanted the string to be *different* than the variable name,
there would have to be a way to do that also, supposing...