+ taskday: no blocked tasks unless the new -b/--blocked provided
normally we can't work on blocked tasks, so there's no use to see them.
this uses a clever (or dumb?) arrangement where any value besides 'None'
is treated as "include one or the other" (corresponding to True and
False), but if it was never set -- relying on the default 'None' -- then
we know to to include both.  this allows 'taskday' to continue its
current status as a "special" caller that can tell us to do special
things by how it sets kwargs (we just used similar thing to implement
'held' flag also)