/ taskadd: send newline on 'n' to avoid error message run-on
when user selects "n"ot to continue, normally the <enter> is not
reproduced so the cursor is not advanced.  thus ensuing error message
runs on to the 'n' keypress itself with no intervening newline

since we know this will happen cause terminal to advance.  we will be
sending to stdout rather than stderr, but that should be fine and have
the same effect as if they user's own <enter> had been echoed back to
the terminal (and stderr from bomb() will begin after the advance)