- taskstop: ignore already-stopped when invoked from on-modify
sometimes we are called to [potentially] stop a job that's already
stopped, in particular when we close a task while a different one is
being worked.  rather than detect this in the trigger we will just make
taskstop ignore already-stopped, but only when called from the hook.

if this happens on the command line, the user would want to know about
it, so we continue to report that.

this is implemented using a new verbose flag which gets set to default
true and this is the case taken from cli invocation, while we override
to false when called from the on-modify hook