^ taskget: remove use of 'abort' kwarg as it was not actually used
the 'abort' kwarg was passed in to taskget from taskone and taskdo, but
not actually used.  taskone already counts its results and aborts if
greater than one (as its name would suggest).  taskdo should be using
taskone rather than setting an ignored abort flag.  probably taskid
should have a flag to stop if more than one match also, and use taskone
if it was given (thus getting a dummy, rather than the first match as it
does today).

the alternative is to support an abort flag in taskget, but it seems
better interface design if taskget always gets everything and outer
functions decide if greater than one result is a problem.