/ taskget: make result list stable when not returning first match
we want taskget results to be deterministic, both when there is one
match and multiple matches.  we were seeing different results from
multiple invocations of 'taskid' because results were not stable.  this
was due to the fact that we use a set to collect and de-dupe objects, so
there is inherently unreliable order.  we can fix this by sorting at the
end, and we'll use task modification time, which is usually what we
would want.  really, we should implement this to use timew order, rather
than taskmod order, but that's a bigger change; at least the result is
now deterministic.

for the 'onlyone' case, ie the first match, we should always get the
same one (ie deterministically) because we store that one specially in
its own variable.  probably we would not want this either if we were
implementing purely by timew sort, but that also has to be thought more
about.

see task 8ae99785