/ tasknow: avoid potentially wrong task when user gave format options
if user specifies a format, we give the fql obtained from _tasknow()
again to _taskids() so that we can pass format kwargs to it and just use
the already-implemented logic in that function to do the printing of the
task for us.  fql is a valid and unique lookup key so this avoids
reimplementing or factoring out any of the logic to print.

however, by giving the fql to _taskids(), we open ourselves to another
possibility: sometimes when an initial substring of an fql is shared with
another task (ie same project and same initial string of label but
different complete label, where first complete label is a subset of
second task label), _taskids() could return multiple results, and could
therefore give us a different result than _tasknow() did (which gets the
real current task, directly from timew data).  we were passing
'firstonly=True' kwarg to make sure to get only one result in the case
that there are several, but this just means we can potentially get a
different value from the list.

instead, we already know we got a complete fql from _tasknow(), so that
string should look up just fine to a single result if given to _taskid()
with idonly=True and exact=True.

this also has the side effect of eliminating one of the two extant users
of 'firstonly' which, as a separate matter, we are working on
eliminating, as part of task 0beaf8e4