- taskids: eliminate 'firstonly' kwarg, taskget already has a 'matchone'
this eliminates a duplicate semantic that required extra logic, by just
passing an already existing kwarg through to _taskget(), which is
already doing what we want with this option.

we also eliminated special-casing for when 'firstonly' was supplied to
_taskids() because there ended up being just two users of this kwarg and
we got right of one of them (tasknow()) in previous commit by switching
from _taskids() to _taskid() which is already going to return only one
result.  for the remaining one (_taskid() itself), we are just passing
in matchone all the way through to _taskget() which already accomplishes
what we need.