+ taskrecent: send to pager, no page if less than screen, and fold lines
we will typically want the output of this paged, but not if we selected
the right timeframe to fit on a single screen.

if run in 80 column terminal, the output will typically be too long, so
we will just truncate that.  in the case that it's more than a screenful
of output, the user will be in a pager and get to see those lines if he
unfolds them or horizontally scrolls