+ taskcont: add taskcont+timecont and tasklast+timelast implementations
sometimes we want to go to the last non-pseudo task (for us this means
"fql not in 'time/' namespace"), and sometimes we want any timew task,
not just taskw-backed.  and sometimes we want the last one, and
sometimes we want the one before last.

so we have 4 permutations here

also we will use this from rptaskline, so it must be invokable as a
script, rather than a shell alias or function.  so we will move it to
.task repo (was in src/.bash)