- taskadd: either show uuid, or start, but not both, helps exit code
if we don't do this exit code is bad in the not-start case, even though
the task was successfully created (see task 2c71aec0)

we could have just always returned success, but having this difference
in behavior is better and collapses to one line.  if the user has
requested to start the newly-added task, they don't care about the id
because starting is already handled for them.  so we can just make this
an ordinary if/then/else