- taskpy: move version check to top and import from sys individually
this is to match better the current devskel implementation.  it's better
when the call names are shorter because they don't need a namespace
qualification.

hexversion check should go at top in case there are things done in the
file which won't even allow it to be read on an earlier python version,
it would crash with a parse error rather than displaying a version
insufficiency message to the user.