- taskget: remove redundant taskfilter call in id test case
this ends up running taskfilter() with same args that it gives to
update_matches() right afterwards.  probably leftover from earlier
implementation.  when running strace we do not see any difference in
execve calls so internally tasklib must cache when it already has seen
the same query.  so this will have no real effect, but makes the code
more clear because we don't have to figure out what this extra filter is
for