/ taskl: simplify with lookup via taskone and restore use of args
args are essential to the purpose of this wrapper because it's supposed
to be the same as typing "task X ...", it just supplies X via taskget
lookup.  so commit 1427955 was a mistake based on misunderstanding
purpose of wrapper.

and to save us some trouble of doing it ourselves, we can use the
already-existing 'taskone' to both (1) guarantee there's only a single
match, and (2) give a bogon uuid on failure, so the wrapped 'task' call
fails.

resulting script ends up being one line