+ tasksh: add name-switched script, all funcs from ~/.bash/func.d/taskw
move these from bash profile so they are easier to modify and catalog,
as part of the taskwtools project

there have been no changes per se.  task() had to be redone to use paths
instead of wrap a function because it's in a script now.  this used some
code from devskel because we do the same thing in other scripts (tmux
wrapper in particular) where it's called the same thing as the thing
it's wrapping and is also a script.  so it just strips its own location
from the path and looks up again.

for older history of these files, see src/.bash@8884894