- timesum: switch back to awk one-liner from before with simpler fix
last time we used this but it doesn't work because timew started
wrapping in columns, so some line would have only wrap content from
previous lines and no data of their own.  this meant suddenly NF == 1
was matching on more than just the summary line at the end.

as a solution we started testing to see if not only NF == 1 but also
it's a time ie digit and ':' only.

but we know something better: the last line is actually the one with the
time summary.  so regardless of wrap it will always be the first NF == 1
if the file is read in reverse line order.

hence, just add a '| tac' and we're back in business